Consolidate duplicate download-repair helpers into core/download_presence.dart
clearDownloadState() (added in #909's core/stale_download.dart, for the
playback path) and resetDownloadState() (this branch's
core/download_presence.dart, for the UI path) were identical - both reset
the same four Episode fields. Per review feedback on #910, folded
everything into a single core/download_presence.dart that both
NativeAudioPlayerService (playback path) and LocalDownloadUtils (UI path)
import, and merged the two test files into one.
